requirements.txt
django==5.0.6
djangorestframework==3.15.1
psycopg2==2.9.9
drf-spectacular==0.27.2
django-filter==24.2
django-debug-toolbar==4.3.0
gunicorn==22.0.0
django-filter==24.2
flake8==7.0.0
black==24.4.2